requirements.txt
pelican>=4.11
markdown
typogrify
pelican-render-math
pelican-neighbors
PyYAML
